+++ SQLiteStudio3/guiSQLiteStudio/dbtree/dbtreeview.h
@@ -31,6 +31,7 @@ class GUI_API_EXPORT DbTreeView : public QTreeView
         DbTreeItem *getItemForAction(bool onlySelected = false) const;
         QPoint getLastDropPosition() const;
         QModelIndexList getSelectedIndexes() const;
+        DbTreeItemDelegate* getItemDelegate() const;
 
     protected:
         void dragEnterEvent(QDragEnterEvent* e);