+++ SQLiteStudio3/guiSQLiteStudio/mainwindow.cpp
@@ -342,7 +342,6 @@ void MainWindow::createActions()
 
     ui->dbToolbar->addAction(dbTree->getAction(DbTree::NEW_DB));
     ui->dbToolbar->addAction(dbTree->getAction(DbTree::OPEN_FILE));
-    ui->dbToolbar->addAction(dbTree->getAction(DbTree::EDIT_DB));
     ui->dbToolbar->addAction(dbTree->getAction(DbTree::REFRESH_SCHEMA));
 
     attachActionInMenu(dbTree->getAction(DbTree::OPEN_FILE), dbTree->getAction(DbTree::ADD_DB), ui->dbToolbar);