+++ SQLiteStudio3/guiSQLiteStudio/windows/tablewindow.cpp
@@ -200,6 +200,9 @@ void TableWindow::init()
     connect(structureExecutor, SIGNAL(success(SqlQueryPtr)), this, SLOT(changesSuccessfullyCommitted()));
     connect(structureExecutor, SIGNAL(failure(int,QString)), this, SLOT(changesFailedToCommit(int,QString)));
 
+    connect(NOTIFY_MANAGER, SIGNAL(objectRenamed(Db*,QString,QString,QString)), this, SLOT(handlePossibleIdxOrTrgRename(Db*,QString,QString,QString)));
+    connect(NOTIFY_MANAGER, SIGNAL(columnRenamed(Db*,QString,QString,QString,QString)), this, SLOT(handlePossibleColumnRename(Db*,QString,QString,QString,QString)));
+
     THEME_TUNER->manageCompactLayout({
                                          ui->structureTab,
                                          ui->constraintsWidget,
@@ -600,7 +603,8 @@ void TableWindow::initDbAndTable()
     ui->triggerList->sortByColumn(0, Qt::AscendingOrder);
 
     // (Re)connect to DB signals
-    connect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
+    if (db)
+        connect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
 
     // Selection model is recreated when setModel() is called on the view
     connect(ui->structureView->selectionModel(), SIGNAL(currentChanged(QModelIndex,QModelIndex)),
@@ -1848,10 +1852,45 @@ void TableWindow::updateFont()
 
 void TableWindow::dbChanged()
 {
-    disconnect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
+    if (db)
+        disconnect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
 
     db = ui->dbCombo->currentDb();
     dataModel->setDb(db);
 
-    connect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
+    if (db)
+        connect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
+}
+
+void TableWindow::handlePossibleIdxOrTrgRename(Db* db, const QString& database, const QString& oldObject, const QString& newObject)
+{
+    Q_UNUSED(database);
+    if (db != this->db)
+        return;
+
+    for (int i = 0, total = ui->indexList->rowCount(); i < total; ++i)
+    {
+        if (ui->indexList->item(i, 0)->text().compare(oldObject, Qt::CaseInsensitive) == 0)
+        {
+            ui->indexList->item(i, 0)->setText(newObject);
+            return;
+        }
+    }
+
+    for (int i = 0, total = ui->triggerList->rowCount(); i < total; ++i)
+    {
+        if (ui->triggerList->item(i, 0)->text().compare(oldObject, Qt::CaseInsensitive) == 0)
+        {
+            ui->triggerList->item(i, 0)->setText(newObject);
+            return;
+        }
+    }
+}
+
+void TableWindow::handlePossibleColumnRename(Db* db, const QString& database, const QString& table, const QString& oldObject, const QString& newObject)
+{
+    if (db != this->db || table.compare(this->table, Qt::CaseInsensitive) != 0)
+        return;
+
+    refreshStructure();
 }